+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [2.0.0]
+### Added
+- hint text and color for comment box
+
 ### Changed
 - **Breaking change:** Replaced "positive/nagative button" clicked listener by a global listener which needs to be implemented by the host activity